[Executorch][llama] Renamed quantized_kv_cache to custom_kv_cache
Pull Request resolved: https://github.com/pytorch/executorch/pull/9944

Because old name was misnomer
ghstack-source-id: 277233486
@exported-using-ghexport

Differential Revision: [D71833067](https://our.internmc.facebook.com/intern/diff/D71833067/)
